+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Journal des modifications
 
 9.3.0
 -----
+- Module Admission : Nouvelle pièce justificative : Attestation de responsabilité civile
 - Module Formation : Possibilité de déposer un fichier PDF pour la fiche de présentation d'une formation
 - Module Formation : Ajout d'un tri pour les individus sur l'écran Inscriptions
 - Module Formation : Ajout d'un bouton sur la fiche d'une session pour envoyer la feuille d'émargement aux formateurs